+++ packages/apps/plugins/plugin-map/src/MapPlugin.tsx
@@ -3,12 +3,14 @@
 //
 
 import { Compass, type IconProps } from '@phosphor-icons/react';
+import { effect } from '@preact/signals-core';
 import React from 'react';
 
-import { SPACE_PLUGIN, SpaceAction } from '@braneframe/plugin-space';
-import { Folder, Map as MapType } from '@braneframe/types';
-import { resolvePlugin, type PluginDefinition, parseIntentPlugin, NavigationAction } from '@dxos/app-framework';
-import { SpaceProxy } from '@dxos/react-client/echo';
+import { parseClientPlugin } from '@braneframe/plugin-client';
+import { updateGraphWithAddObjectAction } from '@braneframe/plugin-space';
+import { Map as MapType } from '@braneframe/types';
+import { resolvePlugin, type PluginDefinition, parseIntentPlugin } from '@dxos/app-framework';
+import { EventSubscriptions } from '@dxos/async';
 
 import { MapMain, MapSection } from './components';
 import meta, { MAP_PLUGIN } from './meta';
@@ -29,35 +31,62 @@ export const MapPlugin = (): PluginDefinition<MapPluginProvides> => {
       },
       translations,
       graph: {
-        builder: ({ parent, plugins }) => {
-          if (!(parent.data instanceof Folder || parent.data instanceof SpaceProxy)) {
+        builder: (plugins, graph) => {
+          const client = resolvePlugin(plugins, parseClientPlugin)?.provides.client;
+          const dispatch = resolvePlugin(plugins, parseIntentPlugin)?.provides.intent.dispatch;
+          if (!client || !dispatch) {
             return;
           }
 
-          const intentPlugin = resolvePlugin(plugins, parseIntentPlugin);
-
-          parent.actionsMap[`${SPACE_PLUGIN}/create`]?.addAction({
-            id: `${MAP_PLUGIN}/create`,
-            label: ['create object label', { ns: MAP_PLUGIN }],
-            icon: (props) => <Compass {...props} />,
-            invoke: () =>
-              intentPlugin?.provides.intent.dispatch([
-                {
+          const subscriptions = new EventSubscriptions();
+          const { unsubscribe } = client.spaces.subscribe((spaces) => {
+            spaces.forEach((space) => {
+              subscriptions.add(
+                updateGraphWithAddObjectAction({
+                  graph,
+                  space,
                   plugin: MAP_PLUGIN,
                   action: MapAction.CREATE,
-                },
-                {
-                  action: SpaceAction.ADD_OBJECT,
-                  data: { target: parent.data },
-                },
-                {
-                  action: NavigationAction.ACTIVATE,
-                },
-              ]),
-            properties: {
-              testId: 'mapPlugin.createObject',
-            },
+                  properties: {
+                    label: ['create object label', { ns: MAP_PLUGIN }],
+                    icon: (props: IconProps) => <Compass {...props} />,
+                    testId: 'mapPlugin.createObject',
+                  },
+                  dispatch,
+                }),
+              );
+
+              // Add all maps to the graph.
+              const query = space.db.query(MapType.filter());
+              let previousObjects: MapType[] = [];
+              subscriptions.add(
+                effect(() => {
+                  const removedObjects = previousObjects.filter((object) => !query.objects.includes(object));
+                  previousObjects = query.objects;
+                  removedObjects.forEach((object) => graph.removeNode(object.id));
+                  query.objects.forEach((object) => {
+                    graph.addNodes({
+                      id: object.id,
+                      data: object,
+                      properties: {
+                        // TODO(wittjosiah): Reconcile with metadata provides.
+                        label: object.title || ['object title placeholder', { ns: MAP_PLUGIN }],
+                        icon: (props: IconProps) => <Compass {...props} />,
+                        testId: 'spacePlugin.object',
+                        persistenceClass: 'echo',
+                        persistenceKey: space?.key.toHex(),
+                      },
+                    });
+                  });
+                }),
+              );
+            });
           });
+
+          return () => {
+            unsubscribe();
+            subscriptions.clear();
+          };
         },
       },
       stack: {