+++ packages/apps/plugins/plugin-script/src/ScriptPlugin.tsx
@@ -3,23 +3,20 @@
 //
 
 import { Code, type IconProps } from '@phosphor-icons/react';
-import React from 'react';
+import { effect } from '@preact/signals-core';
+import React, { useMemo } from 'react';
 
-import { SPACE_PLUGIN, SpaceAction } from '@braneframe/plugin-space';
-import { Folder, Script as ScriptType } from '@braneframe/types';
-import { resolvePlugin, type PluginDefinition, parseIntentPlugin, NavigationAction } from '@dxos/app-framework';
-import {
-  type Filter,
-  type EchoObject,
-  type Schema,
-  TextObject,
-  isTypedObject,
-  SpaceProxy,
-} from '@dxos/react-client/echo';
+import { parseClientPlugin } from '@braneframe/plugin-client';
+import { updateGraphWithAddObjectAction } from '@braneframe/plugin-space';
+import { Script as ScriptType } from '@braneframe/types';
+import { resolvePlugin, type PluginDefinition, parseIntentPlugin } from '@dxos/app-framework';
+import { EventSubscriptions } from '@dxos/async';
+import { createDocAccessor } from '@dxos/react-client/echo';
+import { type Filter, type EchoObject, type Schema, TextObject, isTypedObject } from '@dxos/react-client/echo';
 import { Main } from '@dxos/react-ui';
 import { baseSurface, fixedInsetFlexLayout, topbarBlockPaddingStart } from '@dxos/react-ui-theme';
 
-import { ScriptBlock } from './components';
+import { ScriptBlock, type ScriptBlockProps } from './components';
 import meta, { SCRIPT_PLUGIN } from './meta';
 import translations from './translations';
 import { ScriptAction, type ScriptPluginProvides } from './types';
@@ -49,35 +46,62 @@ export const ScriptPlugin = ({ containerUrl }: ScriptPluginProps): PluginDefinit
       },
       translations,
       graph: {
-        builder: ({ parent, plugins }) => {
-          if (!(parent.data instanceof Folder || parent.data instanceof SpaceProxy)) {
+        builder: (plugins, graph) => {
+          const client = resolvePlugin(plugins, parseClientPlugin)?.provides.client;
+          const dispatch = resolvePlugin(plugins, parseIntentPlugin)?.provides.intent.dispatch;
+          if (!client || !dispatch) {
             return;
           }
 
-          const intentPlugin = resolvePlugin(plugins, parseIntentPlugin);
-
-          parent.actionsMap[`${SPACE_PLUGIN}/create`]?.addAction({
-            id: `${SCRIPT_PLUGIN}/create`,
-            label: ['create object label', { ns: SCRIPT_PLUGIN }],
-            icon: (props) => <Code {...props} />,
-            invoke: () =>
-              intentPlugin?.provides.intent.dispatch([
-                {
+          const subscriptions = new EventSubscriptions();
+          const { unsubscribe } = client.spaces.subscribe((spaces) => {
+            spaces.forEach((space) => {
+              subscriptions.add(
+                updateGraphWithAddObjectAction({
+                  graph,
+                  space,
                   plugin: SCRIPT_PLUGIN,
                   action: ScriptAction.CREATE,
-                },
-                {
-                  action: SpaceAction.ADD_OBJECT,
-                  data: { target: parent.data },
-                },
-                {
-                  action: NavigationAction.ACTIVATE,
-                },
-              ]),
-            properties: {
-              testId: 'scriptPlugin.createObject',
-            },
+                  properties: {
+                    label: ['create object label', { ns: SCRIPT_PLUGIN }],
+                    icon: (props: IconProps) => <Code {...props} />,
+                    testId: 'scriptPlugin.createObject',
+                  },
+                  dispatch,
+                }),
+              );
+
+              // Add all scripts to the graph.
+              const query = space.db.query(ScriptType.filter());
+              let previousObjects: ScriptType[] = [];
+              subscriptions.add(
+                effect(() => {
+                  const removedObjects = previousObjects.filter((object) => !query.objects.includes(object));
+                  previousObjects = query.objects;
+                  removedObjects.forEach((object) => graph.removeNode(object.id));
+                  query.objects.forEach((object) => {
+                    graph.addNodes({
+                      id: object.id,
+                      data: object,
+                      properties: {
+                        // TODO(wittjosiah): Reconcile with metadata provides.
+                        label: object.title || ['object title placeholder', { ns: SCRIPT_PLUGIN }],
+                        icon: (props: IconProps) => <Code {...props} />,
+                        testId: 'spacePlugin.object',
+                        persistenceClass: 'echo',
+                        persistenceKey: space?.key.toHex(),
+                      },
+                    });
+                  });
+                }),
+              );
+            });
           });
+
+          return () => {
+            unsubscribe();
+            subscriptions.clear();
+          };
         },
       },
       stack: {
@@ -100,30 +124,30 @@ export const ScriptPlugin = ({ containerUrl }: ScriptPluginProps): PluginDefinit
             case 'main':
               return isObject(data.active, ScriptType.schema, ScriptType.filter()) ? (
                 <Main.Content classNames={[baseSurface, fixedInsetFlexLayout, topbarBlockPaddingStart]}>
-                  <ScriptBlock
-                    id={(data.active as any).id}
-                    source={(data.active as ScriptType).source}
-                    classes={{ toolbar: 'px-2' }}
+                  <ScriptBlockWrapper
+                    //
+                    script={data.active as ScriptType}
                     containerUrl={containerUrl}
+                    classes={{ toolbar: 'px-1' }}
                   />
                 </Main.Content>
               ) : null;
             case 'slide':
               return isObject(data.slide, ScriptType.schema, ScriptType.filter()) ? (
-                <ScriptBlock
-                  id={(data.slide as any).id}
-                  source={(data.slide as ScriptType).source}
-                  classes={{ root: 'p-24' }}
+                <ScriptBlockWrapper
+                  //
+                  script={data.slide as ScriptType}
+                  containerUrl={containerUrl}
+                  classes={{ toolbar: 'p-24' }}
                   view='preview'
                   hideSelector
-                  containerUrl={containerUrl}
                 />
               ) : null;
             case 'section':
               return isObject(data.object, ScriptType.schema, ScriptType.filter()) ? (
-                <ScriptBlock
-                  id={(data.object as any).id}
-                  source={(data.object as ScriptType).source}
+                <ScriptBlockWrapper
+                  //
+                  script={data.object as ScriptType}
                   containerUrl={containerUrl}
                   classes={{ root: 'h-[400px] py-2' }}
                 />
@@ -144,6 +168,11 @@ export const ScriptPlugin = ({ containerUrl }: ScriptPluginProps): PluginDefinit
       },
     },
   };
+};
+
+const ScriptBlockWrapper = ({ script, ...props }: { script: ScriptType } & Omit<ScriptBlockProps, 'id' | 'source'>) => {
+  const source = useMemo(() => createDocAccessor(script.source), [script.source]);
+  return <ScriptBlock id={script.id} source={source} {...props} />;
 };
 
 // TODO(burdon): Import.