+++ packages/apps/plugins/plugin-search/src/components/SearchMain.tsx
@@ -6,7 +6,7 @@ import React, { type FC, useState } from 'react';
 
 import { useClient } from '@dxos/react-client';
 import { type Space } from '@dxos/react-client/echo';
-import { DensityProvider, useTranslation } from '@dxos/react-ui';
+import { useTranslation } from '@dxos/react-ui';
 import { groupSurface, mx } from '@dxos/react-ui-theme';
 
 import { SearchResults } from './SearchResults';
@@ -33,24 +33,13 @@ export const SearchMain: FC<{ space: Space }> = ({ space }) => {
   };
 
   return (
-    <div className='flex flex-col grow h-full overflow-hidden'>
-      <DensityProvider density='fine'>
-        <div className={mx('flex bs-[--topbar-size] px-2 py-2')}>
-          <Searchbar
-            classes={{ root: 'rounded shadow' }}
-            variant='subdued'
-            placeholder={t('search placeholder')}
-            onChange={setMatch}
-          />
-        </div>
-      </DensityProvider>
+    <div className='flex flex-col grow h-full'>
+      <Searchbar placeholder={t('search placeholder')} onChange={setMatch} />
       {results.length > 0 && (
         <div className='absolute top-[--topbar-size] bottom-0 z-[1]'>
           {/* TODO(burdon): Change to Portal? */}
           <div className={mx('flex flex-col h-full overflow-hidden', groupSurface)}>
-            <DensityProvider density='fine'>
-              <SearchResults items={results} selected={selected} onSelect={handleSelect} />
-            </DensityProvider>
+            <SearchResults items={results} selected={selected} onSelect={handleSelect} />
           </div>
         </div>
       )}