+++ packages/apps/plugins/plugin-template/src/TemplatePlugin.tsx
@@ -2,19 +2,15 @@
 // Copyright 2023 DXOS.org
 //
 
-import { Asterisk, type IconProps } from '@phosphor-icons/react';
+import { Asterisk, Placeholder, type IconProps } from '@phosphor-icons/react';
+import { effect } from '@preact/signals-core';
 import React from 'react';
 
-import { SPACE_PLUGIN, SpaceAction } from '@braneframe/plugin-space';
-import { Folder } from '@braneframe/types';
-import {
-  resolvePlugin,
-  parseIntentPlugin,
-  NavigationAction,
-  type PluginDefinition,
-  definePlugin,
-} from '@dxos/app-framework';
-import { Expando, SpaceProxy } from '@dxos/react-client/echo';
+import { parseClientPlugin } from '@braneframe/plugin-client';
+import { updateGraphWithAddObjectAction } from '@braneframe/plugin-space';
+import { resolvePlugin, parseIntentPlugin, type PluginDefinition } from '@dxos/app-framework';
+import { EventSubscriptions } from '@dxos/async';
+import { Expando, type TypedObject } from '@dxos/react-client/echo';
 
 import { TemplateMain } from './components';
 import meta, { TEMPLATE_PLUGIN } from './meta';
@@ -23,71 +19,99 @@ import { TemplateAction, type TemplatePluginProvides, isObject } from './types';
 
 const typename = 'template'; // Type.schema.typename
 
-export default definePlugin<TemplatePluginProvides>({
-  meta,
-  provides: {
-    metadata: {
-      records: {
-        [typename]: {
-          placeholder: ['object placeholder', { ns: TEMPLATE_PLUGIN }],
-          icon: (props: IconProps) => <Asterisk {...props} />,
+export const TemplatePlugin = (): PluginDefinition<TemplatePluginProvides> => {
+  return {
+    meta,
+    provides: {
+      metadata: {
+        records: {
+          [typename]: {
+            placeholder: ['object placeholder', { ns: TEMPLATE_PLUGIN }],
+            icon: (props: IconProps) => <Asterisk {...props} />,
+          },
+        },
+      },
+      translations,
+      graph: {
+        builder: (plugins, graph) => {
+          const client = resolvePlugin(plugins, parseClientPlugin)?.provides.client;
+          const dispatch = resolvePlugin(plugins, parseIntentPlugin)?.provides.intent.dispatch;
+          if (!client || !dispatch) {
+            return;
+          }
+
+          const subscriptions = new EventSubscriptions();
+          const { unsubscribe } = client.spaces.subscribe((spaces) => {
+            spaces.forEach((space) => {
+              subscriptions.add(
+                updateGraphWithAddObjectAction({
+                  graph,
+                  space,
+                  plugin: TEMPLATE_PLUGIN,
+                  action: TemplateAction.CREATE,
+                  properties: {
+                    label: ['create object label', { ns: TEMPLATE_PLUGIN }],
+                    icon: (props: IconProps) => <Placeholder {...props} />,
+                    testId: 'templatePlugin.createObject',
+                  },
+                  dispatch,
+                }),
+              );
+
+              // Add all documents to the graph.
+              const query = space.db.query({ type: typename });
+              let previousObjects: TypedObject[] = [];
+              subscriptions.add(
+                effect(() => {
+                  const removedObjects = previousObjects.filter((object) => !query.objects.includes(object));
+                  previousObjects = query.objects;
+                  removedObjects.forEach((object) => graph.removeNode(object.id));
+                  query.objects.forEach((object) => {
+                    graph.addNodes({
+                      id: object.id,
+                      data: object,
+                      properties: {
+                        // TODO(wittjosiah): Reconcile with metadata provides.
+                        label: object.title || ['object title placeholder', { ns: TEMPLATE_PLUGIN }],
+                        icon: (props: IconProps) => <Placeholder {...props} />,
+                        testId: 'spacePlugin.object',
+                        persistenceClass: 'echo',
+                        persistenceKey: space?.key.toHex(),
+                      },
+                    });
+                  });
+                }),
+              );
+            });
+          });
+
+          return () => {
+            unsubscribe();
+            subscriptions.clear();
+          };
+        },
+      },
+      surface: {
+        component: ({ data, role }) => {
+          switch (role) {
+            case 'main': {
+              return isObject(data.active) ? <TemplateMain object={data.active} /> : null;
+            }
+          }
+
+          return null;
+        },
+      },
+      intent: {
+        resolver: (intent) => {
+          switch (intent.action) {
+            case TemplateAction.CREATE: {
+              // TODO(burdon): Set typename.
+              return { data: new Expando({ type: 'template' }) };
+            }
+          }
         },
       },
     },
-    translations,
-    graph: {
-      builder: ({ parent, plugins }) => {
-        if (!(parent.data instanceof Folder || parent.data instanceof SpaceProxy)) {
-          return;
-        }
-
-        const intentPlugin = resolvePlugin(plugins, parseIntentPlugin);
-
-        parent.actionsMap[`${SPACE_PLUGIN}/create`]?.addAction({
-          id: `${TEMPLATE_PLUGIN}/create`, // TODO(burdon): Uniformly "create".
-          label: ['create object label', { ns: TEMPLATE_PLUGIN }], // TODO(burdon): "object"
-          icon: (props) => <Asterisk {...props} />,
-          // TODO(burdon): Factor out helper.
-          invoke: () =>
-            intentPlugin?.provides.intent.dispatch([
-              {
-                plugin: TEMPLATE_PLUGIN,
-                action: TemplateAction.CREATE,
-              },
-              {
-                action: SpaceAction.ADD_OBJECT,
-                data: { target: parent.data },
-              },
-              {
-                action: NavigationAction.ACTIVATE,
-              },
-            ]),
-          properties: {
-            testId: 'templatePlugin.createObject',
-          },
-        });
-      },
-    },
-    surface: {
-      component: ({ data, role }) => {
-        switch (role) {
-          case 'main': {
-            return isObject(data.active) ? <TemplateMain object={data.active} /> : null;
-          }
-        }
-
-        return null;
-      },
-    },
-    intent: {
-      resolver: (intent) => {
-        switch (intent.action) {
-          case TemplateAction.CREATE: {
-            // TODO(burdon): Set typename.
-            return { data: new Expando({ type: 'template' }) };
-          }
-        }
-      },
-    },
-  },
-});
+  };
+};