+++ packages/apps/templates/plugin-template/test/scenarios.ts
@@ -9,6 +9,12 @@ import type template from '../src/template.t';
 export const scenarios: InputOf<typeof template>[] = [
   {
     name: 'default',
-    createFolder: true
-  }
+    createFolder: true,
+    defaultPlugins: true,
+  },
+  {
+    name: 'default-bare',
+    createFolder: true,
+    defaultPlugins: false,
+  },
 ];