+++ packages/core/echo/echo-schema/src/object/text-object.ts
@@ -25,6 +25,9 @@ export type AutomergeTextCompat = TypedObject<{
   content?: string;
 }>;
 
+/**
+ * @deprecated
+ */
 export class TextObject extends AbstractEchoObject<TextModel> {
   static [Symbol.hasInstance](instance: any) {
     return !!instance?.[base] && (isActualTextObject(instance) || isAutomergeText(instance));
@@ -36,7 +39,7 @@ export class TextObject extends AbstractEchoObject<TextModel> {
 
     // Redirect to automerge by default.
     if (opts?.automerge ?? true) {
-      const defaultedField = field ?? 'content';
+      const defaultedField = field ?? 'content'; // TODO(burdon): Factor out const.
       return new AutomergeObject(
         {
           kind,