+++ packages/ui/react-ui-mosaic/src/mosaic/DragOverlay.tsx
@@ -31,7 +31,9 @@ export const MosaicDragOverlay = ({ delay = 200, debug = false, ...overlayProps 
   const timer = useRef<ReturnType<typeof setTimeout>>();
 
   useEffect(() => {
-    if (activeItem) {
+    // Only attempt to resolve a foreign overlay tile if the operation is not rejected, since droppables *should* reject
+    // items they don’t recognize and therefore wouldn’t be able to render a tile for. See dxos/dxos#5082.
+    if (activeItem && operation !== 'reject') {
       let container: MosaicContainerProps<any> | undefined;
       let OverlayComponent: MosaicTileComponent<any> | undefined;
       if (overItem?.path) {