+++ packages/ui/react-ui/src/components/Link/Link.tsx
@@ -9,12 +9,16 @@ import React, { type ComponentPropsWithRef, forwardRef } from 'react';
 import { useThemeContext } from '../../hooks';
 import { type ThemedClassName } from '../../util';
 
-export type LinkProps = ThemedClassName<ComponentPropsWithRef<typeof Primitive.a>> & {
-  asChild?: boolean;
-};
+export type LinkProps = ThemedClassName<ComponentPropsWithRef<typeof Primitive.a>> &
+  Partial<{
+    asChild: boolean;
+    variant: 'accent' | 'neutral';
+  }>;
 
-export const Link = forwardRef<HTMLAnchorElement, LinkProps>(({ asChild, classNames, ...props }, forwardedRef) => {
-  const { tx } = useThemeContext();
-  const Root = asChild ? Slot : Primitive.a;
-  return <Root {...props} className={tx('link.root', 'link', {}, classNames)} ref={forwardedRef} />;
-});
+export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
+  ({ asChild, variant, classNames, ...props }, forwardedRef) => {
+    const { tx } = useThemeContext();
+    const Root = asChild ? Slot : Primitive.a;
+    return <Root {...props} className={tx('link.root', 'link', { variant }, classNames)} ref={forwardedRef} />;
+  },
+);